dedupe implementation
